+++ test/t_combiners.jl
@@ -24,6 +24,11 @@ Base.Test.@test all((L |> wcmb) .== L[1:1,:])# the result is identical to the fi
 nbcmb=TL |> j4pr.naivebayescombiner(3)
 Base.Test.@test all((L |> nbcmb) .== [1 1 1 1])
 
+# Test for Naive Bayes combiner - add 2 more label output combinations 
+#to L that were seen in training 
+bkscmb=TL |> j4pr.bkscombiner(3) 
+Base.Test.@test all(([L [1,1,1] [2,0,1]] |> bkscmb) .== [0 0 0 0 1 2])
+
 # Test for the mean combiner (3 outputs, 1 continuous output)
 Base.Test.@test L |> j4pr.meancombiner(3,1) ≈ [2/3 1/3 4/3 1.0]