+++ Changelog.md
@@ -5,6 +5,7 @@
   * Wrapped Factor Analysis (requires latestMultivariateStats.jl)
   * Wrapped kernel PCA (requires latestMultivariateStats.jl)
   * Added decision stump classifier, regressor
+  * Added BKS label combiner
 
 0.1.0 / 2017-09-24
 ==================